new API

+++ src/services/order.js
@@ -1,80 +1,48 @@
 const orderModel = require('../models/orderModel')
 const clientModel = require('../models/clientModel')
+const productModel = require('../models/productModel')
 const { v4: uuidv4 } = require('uuid')
-
-function validateDate(d){
-  var data = d;
-  data = data.replace(/\//g, "/");
-  var data_array = data.split("/");
-  var dia = data_array[2];
-  var mes = data_array[1];
-  var ano = data_array[0];
-
-  if(data_array[0].length !== 4){
-     dia = data_array[0];
-     mes = data_array[1];
-     ano = data_array[2];
-  }
-
-  var hoje = new Date();
-  var d1 = hoje.getDate();
-  var m1 = hoje.getMonth()+1;
-  var a1 = hoje.getFullYear();
-
-  d1 = new Date(a1, m1, d1);
-  var d2 = new Date(ano, mes, dia);
-
-  var diff = d2.getTime() - d1.getTime();
-  diff = diff / (1000 * 60 * 60 * 24);
-  
-  if(diff < 0){
-    return false
-  } else {
-    return true
-  }
-}
 
 class Order {
   async create (req, res, next) {
     try {
-      const { deliveryDate, orderItems, formOfPayment, clientId, value } = req.body
+      const { clientId, request  } = req.body
 
-      if (!validateDate(deliveryDate)) throw({ _message: "A data de entrega não pode ser anterior ao dia atual."})
-      
-
-      const client = await clientModel.findOne({ clientId })
-      const status = "Aguardando resposta"
+      const { name, phone, address } = await clientModel.findOne({ clientId })
       
       const orderId = uuidv4()
+      
+      const formatedRequest = await Promise.all(request.products.map(async (productId) => {
+        const product = await productModel.findOne({ productId })
+        return product
+      }))
 
-      let orderCreationDate = `16/06/2021`
+      let value = 0
+      for (let product of formatedRequest) {
+        value = value + product.price;
+      }
 
       const order = await orderModel.create({
         orderId,
-        orderCreationDate,
-        deliveryDate,
-        orderItems,
-        status,
-        formOfPayment,
-        value,
-        clientId,
-        clientName: client.name
+        client: {
+          clientId,
+          name,
+          phone,
+          address,
+        },
+        request: {
+          products: formatedRequest,
+          observation: request.observation,
+          amount: value,
+        }
       })
 
       res.status(201).send({
-        orderId: order.orderId,
-        orderCreationDate: order.orderCreationDate,
-        deliveryDate: order.deliveryDate,
-        orderItems: order.orderItems,
-        status: order.status,
-        formOfPayment: order.formOfPayment,
-        value: order.value,
-        clientId: order.clientId,
-        clientName: order.clientName,
+        orderId,
+        client: order.client,
+        request: order.request,
+        amount: order.amount
       })
-
-      res.status(200).send()
-
     } catch (error) {
       if(error._message === 'Order validation failed') {
         error._message = "Há campos em branco"
@@ -93,14 +61,9 @@ class Order {
       res.status(200).json(orders.map(order => (
         {
           orderId: order.orderId,
-          orderCreationDate: order.orderCreationDate,
-          deliveryDate: order.deliveryDate,
-          orderItems: order.orderItems,
-          customization: order.customization,
-          status: order.status,
-          formOfPayment: order.formOfPayment,
-          clientId: order.clientId,
-          clientName: order.clientName
+          client: order.client,
+          request: order.request,
+          amount: order.amount,
         }
       )))
     } catch (error) {
@@ -118,33 +81,11 @@ class Order {
 
       if (!order) return res.status(404).send({error: "Order not found."})
     
-      res.status(200).json({ 
-        orderId: order.orderId, 
-        orderCreationDate: order.orderCreationDate,
-        deliveryDate: order.deliveryDate,
-        orderItems: order.orderItems,
-        status: order.status,
-        formOfPayment: order.formOfPayment,
-        clientId: order.clientId,
-        clientName: order.clientName,
-        value: order.value
-      })
-    } catch (error) {
-      res.status(400)
-      next(error)
-    }
-  }
-
-  async getByOrderIdStatus (req, res, next) {
-    try {
-      const { orderId } = req.params
-
-      const order = await orderModel.findOne({ orderId })
-
-      if (!order) return res.status(404).send({error: "Order not found."})
-    
-      res.status(200).json({ 
-        status: order.status,
+      res.status(200).json({
+        orderId: order.orderId,
+        client: order.client,
+        request: order.request,
+        amount: order.amount,
       })
     } catch (error) {
       res.status(400)
@@ -166,13 +107,8 @@ class Order {
 
         res.status(200).json({
           orderId: response.orderId,
-          orderCreationDate: response.orderCreationDate,
-          deliveryDate: response.deliveryDate,
-          orderItems: response.orderItems,
-          customization: response.customization,
-          status: response.status,
-          formOfPayment: response.formOfPayment,
-          clientId: response.clientId
+          client: response.client,
+          request: response.request,
         })
       } else {
         res.status(400).send({error: "Unable to update order."})
@@ -186,6 +122,9 @@ class Order {
   async deleteByOrderId (req, res, next) {
     try {
       const { orderId } = req.params
+      const orderExists = await orderModel.findOne({ orderId })
+
+      if (!orderExists) return res.status(400).send({error: "Order not found."})
 
       await orderModel.deleteOne({ orderId })
       res.status(200).send({message: "Order deleted."})
